#8 | Extract logic to internal package
