+++ internal/const.go
@@ -1,0 +1,7 @@
+package internal
+
+const (
+	FlagName = "test.v"
+	Lookup   = "go.mod"
+	RootPath = "/"
+)
